requirements.txt
Django==1.9.7
django-autocomplete-light==3.1.6
django-ckeditor==5.0.3
django-import-export==0.4.5
django-multiselectfield==0.1.3
git+git://github.com/mariocesar/sorl-thumbnail@v12.3#egg=sorl-thumbnail
django-markdownx==1.6
WeasyPrint==0.30
django-pipeline==1.6.9